+++ toucan-shopping-modules/user/toucan-shopping-user-entity-vo/src/main/java/com/toucan/shopping/modules/user/vo/freightTemplate/UBCIFreightTemplateDefaultRuleVO.java
@@ -5,6 +5,8 @@ import com.alibaba.fastjson.serializer.ToStringSerializer;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 运费模板默认规则
  *
@@ -21,13 +23,13 @@ public class UBCIFreightTemplateDefaultRuleVO {
 
     private String transportModel; //运送方式 1:快递 2:EMS 3:平邮
 
-    private Double defaultWeight; //默认XXX.XX重量以内 单位KG/件
+    private BigDecimal defaultWeight; //默认XXX.XX重量以内 单位KG/件
 
-    private Double defaultWeightMoney; //默认XXX.XX重量以内金额
+    private BigDecimal defaultWeightMoney; //默认XXX.XX重量以内金额
 
-    private Double defaultAppendWeight; //默认增加XXX.XX重量以内 单位KG/件
+    private BigDecimal defaultAppendWeight; //默认增加XXX.XX重量以内 单位KG/件
 
-    private Double defaultAppendWeightMoney; //默认增加XXX.XX重量以内金额
+    private BigDecimal defaultAppendWeightMoney; //默认增加XXX.XX重量以内金额
